+++ packages/eyes-selenium/src/spec-driver.ts
@@ -45,10 +45,11 @@ export function isSelector(selector: any): selector is Selector {
   )
 }
 export function transformDriver(driver: Driver): Driver {
+  driver.getExecutor().defineCommand('getSessionDetails', 'GET', '/session/:sessionId')
+  driver.getExecutor().defineCommand('getOrientation', 'GET', '/session/:sessionId/orientation')
+  driver.getExecutor().defineCommand('performTouch', 'POST', '/session/:sessionId/touch/perform')
   if (process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION === '3') {
-    const cmd = require('selenium-webdriver/lib/command')
-    cmd.Name.SWITCH_TO_PARENT_FRAME = 'switchToParentFrame'
-    driver.getExecutor().defineCommand(cmd.Name.SWITCH_TO_PARENT_FRAME, 'POST', '/session/:sessionId/frame/parent')
+    driver.getExecutor().defineCommand('switchToParentFrame', 'POST', '/session/:sessionId/frame/parent')
   }
   return driver
 }
@@ -77,8 +78,8 @@ export async function mainContext(driver: Driver): Promise<Driver> {
 }
 export async function parentContext(driver: Driver): Promise<Driver> {
   if (process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION === '3') {
-    const cmd = require('selenium-webdriver/lib/command')
-    await (driver as any).schedule(new cmd.Command(cmd.Name.SWITCH_TO_PARENT_FRAME))
+    const {Command} = require('selenium-webdriver/lib/command')
+    await (driver as any).schedule(new Command('switchToParentFrame'))
     return driver
   }
   await driver.switchTo().parentFrame()
@@ -99,20 +100,14 @@ export async function findElement(driver: Driver, selector: Selector): Promise<E
 export async function findElements(driver: Driver, selector: Selector): Promise<Element[]> {
   return driver.findElements(transformSelector(selector))
 }
-export async function getElementRect(
-  _driver: Driver,
-  element: Element,
-): Promise<{x: number; y: number; width: number; height: number}> {
-  return element.getRect()
-}
 export async function getWindowSize(driver: Driver): Promise<{width: number; height: number}> {
   try {
     const window = driver.manage().window()
-    if (utils.types.isFunction(window.getRect)) {
+    if (utils.types.isFunction(window.getSize)) {
+      return await window.getSize()
+    } else if (utils.types.isFunction(window.getRect)) {
       const rect = await window.getRect()
       return {width: rect.width, height: rect.height}
-    } else if (utils.types.isFunction(window.getSize)) {
-      return await window.getSize()
     }
   } catch (err) {
     // workaround for Appium
@@ -129,34 +124,49 @@ export async function setWindowSize(driver: Driver, size: {width: number; height
     await window.setSize(size.width, size.height)
   }
 }
-export async function getOrientation(driver: Driver): Promise<'landscape' | 'portrait'> {
-  const capabilities = await driver.getCapabilities()
-  const orientation = capabilities.get('orientation') || capabilities.get('deviceOrientation')
-  return orientation.toLowerCase()
-}
 export async function getDriverInfo(driver: Driver): Promise<any> {
+  const session = await driver.getSession()
   const capabilities = await driver.getCapabilities()
   const desiredCapabilities = capabilities.get('desired') ?? {}
-  const session = await driver.getSession()
-  const sessionId = session.getId()
-  const deviceName = desiredCapabilities.deviceName ?? capabilities.get('deviceName')
   const platformName =
     capabilities.get('platformName') ?? capabilities.get('platform') ?? desiredCapabilities.platformName
-  const platformVersion = capabilities.get('platformVersion')
-  const browserName = capabilities.get('browserName') ?? desiredCapabilities.browserName
-  const browserVersion = capabilities.get('browserVersion') ?? capabilities.get('version')
   const isMobile = ['android', 'ios'].includes(platformName?.toLowerCase())
 
-  return {
-    sessionId,
+  const info: any = {
+    sessionId: session.getId(),
     isMobile,
-    isNative: isMobile && !browserName,
-    deviceName,
+    isNative: isMobile && !capabilities.get('browserName'),
+    deviceName: desiredCapabilities.deviceName ?? capabilities.get('deviceName'),
     platformName,
-    platformVersion,
-    browserName,
-    browserVersion,
+    platformVersion: capabilities.get('platformVersion'),
+    browserName: capabilities.get('browserName') ?? desiredCapabilities.browserName,
+    browserVersion: capabilities.get('browserVersion') ?? capabilities.get('version'),
   }
+
+  if (info.isNative) {
+    let details
+    if (capabilities.has('viewportRect') && capabilities.has('pixelRatio')) {
+      details = {viewportRect: capabilities.get('viewportRect'), pixelRatio: capabilities.get('pixelRatio')}
+    } else {
+      const {Command} = require('selenium-webdriver/lib/command')
+      const getSessionDetailsCommand = new Command('getSessionDetails')
+      details =
+        process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION === '3'
+          ? await (driver as any).schedule(getSessionDetailsCommand)
+          : await driver.execute(getSessionDetailsCommand)
+    }
+
+    info.pixelRatio = details.pixelRatio
+    if (details.viewportRect) {
+      info.viewportRegion = {
+        x: details.viewportRect.left,
+        y: details.viewportRect.top,
+        width: details.viewportRect.width,
+        height: details.viewportRect.height,
+      }
+    }
+  }
+  return info
 }
 export async function getTitle(driver: Driver): Promise<string> {
   return driver.getTitle()
@@ -200,6 +210,43 @@ export async function waitUntilDisplayed(driver: Driver, selector: Selector, tim
 
 // #endregion
 
+// #region MOBILE COMMANDS
+
+export async function getOrientation(driver: Driver): Promise<'portrait' | 'landscape'> {
+  const {Command} = require('selenium-webdriver/lib/command')
+  const getOrientationCommand = new Command('getOrientation')
+  const orientation =
+    process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION === '3'
+      ? await (driver as any).schedule(getOrientationCommand)
+      : await driver.execute(getOrientationCommand)
+  return orientation.toLowerCase() as 'portrait' | 'landscape'
+}
+export async function getElementRegion(
+  _driver: Driver,
+  element: Element,
+): Promise<{x: number; y: number; width: number; height: number}> {
+  return element.getRect()
+}
+export async function getElementAttribute(_driver: Driver, element: Element, attr: string): Promise<string> {
+  return element.getAttribute(attr)
+}
+export async function getElementText(_driver: Driver, element: Element): Promise<string> {
+  return element.getText()
+}
+export async function performAction(driver: Driver, steps: any[]): Promise<void> {
+  const {Command} = require('selenium-webdriver/lib/command')
+  const performTouchCommand = new Command('performTouch').setParameters({
+    actions: steps.map(({action, ...options}) => ({action, options})),
+  })
+  if (process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION === '3') {
+    await (driver as any).schedule(performTouchCommand)
+  } else {
+    await driver.execute(performTouchCommand)
+  }
+}
+
+// #endregion
+
 // #region TESTING
 
 const browserOptionsNames: Record<string, string> = {
@@ -211,7 +258,7 @@ export async function build(env: any): Promise<[Driver, () => Promise<void>]> {
   const parseEnv = require('@applitools/test-utils/src/parse-env')
 
   const {
-    browser = '',
+    browser,
     capabilities,
     url,
     attach,
@@ -221,7 +268,7 @@ export async function build(env: any): Promise<[Driver, () => Promise<void>]> {
     args = [],
     headless,
   } = parseEnv({...env, legacy: env.legacy ?? process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION === '3'})
-  const desiredCapabilities = {browserName: browser, ...capabilities}
+  const desiredCapabilities = {...capabilities}
   if (configurable) {
     const browserOptionsName = browserOptionsNames[browser || desiredCapabilities.browserName]
     if (browserOptionsName) {